pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devdox-ai-context"
version = "0.0.1"
description = "Queue Worker Service for DevDox AI Context Processing"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "asyncio",
    "asyncpg==0.29.0",
    "tortoise-orm==0.24.0",
    "pydantic==2.8.0",
    "pydantic-settings==2.5.2",
    "supabase==2.16.0",
    "redis==5.0.1",
    "tembo-pgmq-python==0.10.0",
    "gitpython==3.1.44",
    "pygithub==2.6.1",
    "requests==2.32.4",
    "python-gitlab==4.2.0",
    "structlog==23.2.0",
    "prometheus-client==0.19.0",
    "httpx==0.28.1",
    "dependency-injector==4.48.1",
    "cryptography==44.0.3",
    "aerich==0.9.1",
    "langchain==0.3.13",
    "langchain-community==0.3.13",
    "langchain-mongodb==0.6.1",
    "langchain-together==0.3.0",
    "together==1.5.17",
    "models @ git+https://github.com/montymobile1/devdox-ai-models.git@c7b72e1db2fa646a94e9209d2818e3f55a26cc28",
    "encryption-service @ git+https://github.com/montymobile1/devdox-ai-encryption.git4eefae1c5dfdc6a2ecbdf9acdcb4c099db8523e1"

]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
    "bandit>=1.7.5",
]

[project.scripts]
devdox-worker = "app.main:main"
devdox-context-worker = "app.worker:start_worker"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "models*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=app --cov-report=html --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app", "models"]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
